# backend/go/services/tenant-manager/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreateTenantRequest:
    properties:
      billing_email:
        type: string
      email:
        type: string
      metadata:
        properties:
          company_size:
            type: string
          industry:
            type: string
          website:
            type: string
        type: object
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        type: string
      plan:
        enum:
        - starter
        - professional
        - enterprise
        type: string
    required:
    - email
    - name
    - plan
    type: object
  handler.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  handler.IncrementUsageRequest:
    properties:
      calls:
        type: integer
      minutes:
        type: integer
    type: object
  handler.ListTenantsResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      tenants:
        items:
          $ref: '#/definitions/handler.TenantResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handler.QuotaResponse:
    properties:
      max_api_keys:
        type: integer
      max_calls_per_month:
        type: integer
      max_minutes_per_month:
        type: integer
      max_storage_gb:
        type: integer
      max_users:
        type: integer
      reset_at:
        type: string
      tenant_id:
        type: string
      used_calls:
        type: integer
      used_minutes:
        type: integer
      used_storage_gb:
        type: number
    type: object
  handler.TenantResponse:
    properties:
      billing_email:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      phone:
        type: string
      plan:
        type: string
      settings:
        additionalProperties: true
        type: object
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  handler.UpdateQuotaRequest:
    properties:
      max_api_keys:
        type: integer
      max_calls_per_month:
        type: integer
      max_minutes_per_month:
        type: integer
      max_storage_gb:
        type: integer
      max_users:
        type: integer
      reset_at:
        type: string
    type: object
  handler.UpdateTenantRequest:
    properties:
      billing_email:
        type: string
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        type: string
    type: object
info:
  contact: {}
  description: Tenant and API key management service
  title: Tenant Manager API
  version: "1.0"
paths:
  /api/v1/tenants:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Tenant status filter
        in: query
        name: status
        type: string
      - description: Search by name or email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListTenantsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tenants
      tags:
      - Tenants
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create tenant
      tags:
      - Tenants
  /api/v1/tenants/{id}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete tenant
      tags:
      - Tenants
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tenant
      tags:
      - Tenants
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update tenant
      tags:
      - Tenants
...
  /api/v1/tenants/{id}/quota:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.QuotaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tenant quota
      tags:
      - Tenants
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Quota payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateQuotaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.QuotaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update tenant quota
      tags:
      - Tenants
  /api/v1/tenants/{id}/usage:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Usage payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.IncrementUsageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.QuotaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Increment tenant usage
      tags:
      - Tenants
securityDefinitions:
  BearerAuth:
    description: 'Bearer access token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"